requirements.txt
gradio==5.25.0
ecologits==0.6.1

litellm==1.68.1

# For db export, could be in dev image
SQLAlchemy==2.0.38

google-cloud-aiplatform
huggingface-hub
fastapi
tomli
json5
gradio_modal
openai
sentry_sdk
google-auth
pandas
psycopg2-binary
tiktoken
langfuse
pyarrow

# Docs authentication dependencies (same as Drive)
httpx
python-jose[cryptography]